backend/pyproject.toml
[project]
name = "seer-ai-backend"
version = "0.1.0"
description = "FRC Robot and Ball Tracking Backend"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "yt-dlp>=2024.11.0",
    "ultralytics>=8.3.0",
    "opencv-python-headless>=4.10.0",
    "numpy>=1.26.0",
    "matplotlib>=3.9.0",
    "scipy>=1.14.0",
    "python-multipart>=0.0.12",
    "aiofiles>=24.1.0",
    "Pillow>=10.4.0",
    "seaborn>=0.13.0",
    "websockets>=12.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.uv]
dev-dependencies = []
